(ERL) Fix global namespace for static libraries.
This bug was here since the initial import. The linker usually only
pulls directly referenced symbols, but if -Wl,--whole-archive is used
then the ERL metadata becomes visible creating conflicts. The erl-loader
doesn't add the extra symbols to the global namespace when loading ERLs
so they should only be in the ERL archives.

I also replaced the use of dependancy et al, with dependency.